pyproject.toml
[tool.poetry]
name = "tvscraper"
version = "1.0.0"
description = "Extract watch list from TradingView using Selenium"
authors = ["Ethan Ho <eho@tacc.utexas.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pytest-pythonpath = "^0.7.3"
selenium = "^3.141.0"
attrs = "^20.3.0"
attrdict = "^2.0.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
python-dotenv = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"